converted most of the style to tailwind, remove useless, old course payment modal
