+++ apps/web/public/locales/en.json
@@ -161,7 +161,7 @@
       "by": "by {{professor}}",
       "letsgo": "Let's go"
     },
-    "purchase": "Purchase",
+    "proceedToPayment": "Proceed to payment",
     "quizz": {
       "clickOnPie": "Click on the pie chart, so check your mistakes, and see the explanations",
       "quizz": "Quizz",
